Style Changes
Added various stlying changes to the quiz using the jumbotron for the header and then having containers with bootstrap backgrounds added to them to attempt make them more appealing to the eye. Added spacing to the question containers and rounded off the edges. Added a Submit Quiz button to the bottom.